Fixes context generation for non-chat providers.
When we're calling provider.generate, ultimately we're calling the generate
 function as specified in langchain_core.language_models.llms.BaseLLM.
This function expects a list of string prompts, where each should have a
separate generation performed for it - this being to allow taking advantage
 of batch processing where possible.

The bug here is that we pass in each message from the context as a separate
generation question,  rather than combining them together into a single
prompt including both the context and human message.

The edit here concatenates them all with two new lines between each message
 and ensures that we pass only a single element list to the generate
function.