Fix tool call UI by renaming id to tool_id
The id attribute is a reserved HTML attribute and was not being properly
passed to the React component when used in web components. This caused
the tool call UI to not display because the component returned null when
props.id was undefined.

Changed all instances of id to tool_id in:
- toolcall_list.py: Changed prop key from "id" to "tool_id"
- types.py: Updated JaiToolCallProps type definition
- web-components-plugin.ts: Updated prop declaration and allowed attributes
- jai-tool-call.tsx: Updated TypeScript interface and usage